# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de Agenda",
        "description": "L5 Swagger",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tareas": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Listar todas las tareas del usuario autenticado",
                "operationId": "6733ab2dc67122575c368d93ff4c1308",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado de la tarea (pendiente/completada)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha",
                        "in": "query",
                        "description": "Filtrar por fecha límite de la tarea",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tarea"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "cff34cbb3cdf575de33470453733f08c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Comprar pan"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Ir a la panadería a comprar pan"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-30"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tarea"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tareas/{id}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Ver detalle de una tarea",
                "operationId": "e67767d5fb94e0ee2463e80560e4c7e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la tarea",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tarea"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualizar una tarea",
                "operationId": "883b7dcb1795b74253c21dd6aa80507c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Comprar pan actualizado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Ir a la panadería y comprar pan y leche"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-30"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "completada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tarea"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "8c3e8a0e10b85427ca346b93d3bcc223",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarea eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tarea": {
                "title": "Tarea",
                "properties": {
                    "id": {
                        "description": "* @OA\\Schema(\n*     schema=\"Tarea\",\n*     type=\"object\",\n*     title=\"Tarea\",\n*     @OA\\Property(property=\"id\", type=\"integer\", example=1),\n*     @OA\\Property(property=\"user_id\", type=\"integer\", example=1),\n*     @OA\\Property(property=\"titulo\", type=\"string\", example=\"Comprar pan\"),\n*     @OA\\Property(property=\"descripcion\", type=\"string\", example=\"Ir a la panadería a comprar pan\"),\n*     @OA\\Property(property=\"fecha_limite\", type=\"string\", format=\"date\", example=\"2025-11-30\"),\n*     @OA\\Property(property=\"estado\", type=\"string\", example=\"pendiente\"),\n*     @OA\\Property(property=\"created_at\", type=\"string\", format=\"date-time\"),\n*     @OA\\Property(property=\"updated_at\", type=\"string\", format=\"date-time\")\n* )",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Comprar pan"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Ir a la panadería a comprar pan"
                    },
                    "fecha_limite": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-30"
                    },
                    "estado": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}